Rename sign_on to admin_ui
- Ensure that everything relating to the admin UI is in one file
- This should (?) make it easier to extract to a separate project in future
